80-85% complete
Added a science page. I will add to it w/ actually facts and links to articles later. Most of the main elements (pictures, inputs, navbar) have been coded in. Still need to add buttons to the resource page. The main focus for future commits and edits to this project will be about CSS styling.